Calculate stochastic epi mutations from a methylation dataset as outcome report of pivot

Calculate stochastic epi mutations from a methylation dataset as outcome report of pivot

analize_population(
  methylation_data,
  sliding_window_size,
  beta_superior_thresholds,
  beta_inferior_thresholds,
  sample_sheet,
  beta_medians,
  bonferroni_threshold = 0.05,
  probe_features
)

Arguments

methylation_data

whole matrix of data to analyze.

sliding_window_size

size of the sliding widows to compute epilesions default 11 probes.

beta_superior_thresholds

data frame to select, from the sample sheet, samples to use as control as study population and as refereces two vectors within the first vector the names of the selection colum and tha second vector the study population selector,

beta_inferior_thresholds

name of samplesheet's column to use as control population selector followed by selection value,

sample_sheet

name of samplesheet's column to use as control population selector followed by selection value,

beta_medians

name of samplesheet's column to use as control population selector followed by selection value

bonferroni_threshold

threshold to define which pValue accept for

probe_features

probes detail from 27 to EPIC illumina dataset lesions definition

Value

files into the result folder with pivot table and bedgraph.