takes a bed and its location (build with the details of popuationa nd genomic area) and annoate with detail about genomic area

takes a bed and its location (build with the details of popuationa nd genomic area) and annoate with detail about genomic area

annotate_bed(
  envir,
  populations,
  figures,
  anomalies,
  groups,
  probes_prefix,
  columnLabel,
  groupingColumnLabel
)

Arguments

envir

semseekere working infos

populations

vector of population to cycle with to build the folder path

figures

vector of hyper /hypo to use to build the folder path

anomalies

vector of lesions/mutations to use to build the folder path

groups

vector of genomic area to cycle and group the annotated data

probes_prefix

prefix to use to get the annotated probes dataset

columnLabel

label of the column of the genomic area gene, island ,dmr etc..

groupingColumnLabel

label of the column of the genomic sub area body, tss1500

Value

original bed with genomic area infos